tests/wpt/tests/css/css-scroll-snap/parsing/scroll-padding-shorthand.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Scroll Snap Test: scroll-padding sets longhands</title>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1/#propdef-scroll-padding">
<meta name="assert" content="scroll-padding supports the full grammar '[ <length-percentage> | auto ]{1,4}'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/shorthand-testcommon.js"></script>
</head>
<body>
<script>
test_shorthand_value('scroll-padding', '10px', {
  'scroll-padding-top': '10px',
  'scroll-padding-right': '10px',
  'scroll-padding-bottom': '10px',
  'scroll-padding-left': '10px'
});

test_shorthand_value('scroll-padding', 'auto 20px', {
  'scroll-padding-top': 'auto',
  'scroll-padding-right': '20px',
  'scroll-padding-bottom': 'auto',
  'scroll-padding-left': '20px'
});

test_shorthand_value('scroll-padding', '1px 2px 3px', {
  'scroll-padding-top': '1px',
  'scroll-padding-right': '2px',
  'scroll-padding-bottom': '3px',
  'scroll-padding-left': '2px'
});

test_shorthand_value('scroll-padding', '1% 2px 3px 0', {
  'scroll-padding-top': '1%',
  'scroll-padding-right': '2px',
  'scroll-padding-bottom': '3px',
  'scroll-padding-left': '0px'
});
</script>
</body>
</html>
